refactor: extract frontend unit tests into _tests.rs companion files
